total size for active=true
